Switch MacroExpansionContext to an protocol
The macros design has moved to having `MacroExpansionContext` be a
class-bound protocol. Adjust the implementation accordingly, and add a
`TestingMacroExpansionContext` for a concrete implementation to use in
test cases.

And since I didn't want to break the `#fileID` macro that used to work,
implement a general scheme for getting the location of a syntax node :).