[Parser] Go back to parsing macro expansions as expressions where possible
The change to eagerly parsing macro expansions as declarations rather
than expressions caused a regression for expressions that start with a
macro expansion, e.g., `#fileID == "Module/filename.swift"`. Rather
than perform extensive lookahead to disambiguate, go back to parsing
as a macro expansion expression except in places where only a
declaration is allowed (e.g., a member list).

To accommodate this for freestanding macro implementations, generalize
the `CodeItemMacro` and `DeclarationMacro` expansion operations to
deal with an arbitrary `FreestandingMacroExpansionSyntax`, as we had
already done for `ExpressionMacro`.

Fixes #1292.