Remove trailing whitespace in multi-line string literals of tests
In mulit-line string literal tests these trailing whitespaces are really hard to spot but influence the test significantly. To make them more obvious, use `\u{20}`, which is visible in the editor and will be replaced by a normal space in `AssertParse`

In all other cases where the trailing whitespace doesn’t matter, remove it.