[reference-binding] Fix an ambiguity when parsing inout in function types.
This was exposed by:

decl/enum/enumtest.swift

The problem here is that we are now allowing for inout to be the start of a
swift decl. This means that if one defines an enum case with a inout parameter,
the parser gets confused and doesnt think it is part of the type. I worked
around this by changing canParseTupleTypeBody to say that an inout binding
cannot be parsed as part of a tuple type. This fits already with how we want to
model this.