Complete regex literal lexing
The previous regex literal lexing logic didn't
implement all the heuristics needed to parse all
the regex literals that the old C++ parser handles.
Update it with the heuristics it needs, and better
handle diagnostics for invalid cases.

This unfortunately requires a bit of lexical
hackery, including a heuristic to classify previous
token kinds to determine if we're in regex literal
position, but it's needed to handle pathological
cases such as `x /^ y/` where there's an ambiguity
with a binary operator.