[CMake] Updates to allow inclusion using FetchContent
Various updates that allow swift-syntax to be included using
FetchContent.

Mostly this is just not replacing `target_link_libraries` since that is
a global replacement, but we also now use a couple of variables from the
swift if they're set (eg. `SWIFT_HOST_LIBRARIES_DEST_DIR` and
`SWIFT_HOST_MODULE_TRIPLE`).