Add infrastructure for automated validation of syntax node structure
The goal here is to generalize the API inconsistency reported by @AnthonyLatsis and automatically detect them.

There are valid reasons for nodes or children to violate these base rules. My idea is that we note these exceptions in expected failures, ideally with a sort explanation that motivates the deviation from standard naming rules.

We aren’t running these tests at the moment yet. I’d like to merge them as a starting point for API improvement ideas at the moment.

The current rules are
- All nodes with base kind e.g. `ExprSyntax` should end with `ExprSyntax`.
- If a child has a single keyword as its choice, it should be named `*Keyword` (e.g. `ImportKeyword`)
- If a child has a single non-keyword token kind, name it the same as the the token choice (e.g. `LeftParen`)
- If a token only has keyword token choices, its name should end with `Keyword`.
- Children that have the same type also have the same child name.
- Every node that can conform to a trait does so