Support construction of `DeclSyntax` subtypes from string interpolation
Allow the creation of e.g. an `ExtensionDeclSyntax` as

```swift
let extension = try ExtensionDeclSyntax("extension Foo: Proto {}")
```

The initializer throws if the string describes a declaration other than an 'extension'. If the string can get parsed as an extension but contains syntax errors, we return a node that contains syntax errors. This matches the behavior of the `HasTrailingCodeBlock` intializers that take a header + result builder.