Make `SyntaxCollection` conform to `ExpressibleByArrayLiteral`
This hoists the ability to create syntax collections from `SwiftSyntaxBuilder` to the `SwiftSyntax` module.

Impact on API compatibility: Initializing an `ExprListSyntax` or `UnexpectedNodesSyntax` now requires the nodes to be `ExprSyntax` and `Syntax` opposed to `ExprSyntaxProtocol` or `SyntaxProtocol`, respectively. `ExprListSyntax` is used as the child of a `SequenceExprSyntax`. I think this is OK because
- Users don’t genrally construct `UnexpectedNodesSyntax`
- To construct a `SequenceExprSyntax`, you can use the result builder on `SequenceExprSyntax`, which takes `ExprSyntaxProtocol`