Add a release notes document for swift-syntax 510
My idea is that we can build up the list of changes during development. That will make it more transparent for adopters of the snapshot tags to see what has changed and, in case of deprecations or API-incompatible changes, how they need to update.

It will also make it easier to compose the list of changes for the proper (non-snapshot) release of swift-syntax.