Set the privacy level of string interpolation errors in os_log to private
If string interpolation results in parsing errors, we create an os_log fault to notify the user about a potential bug. This log message contains the source code with the syntax error, which could be sensitive.

Set the os_log privacy level to `private` to make sure the source code does not get persisted in logs. It will still show up in the Xcode console.