Tweak grouped-diagnostics formatting to put the primary diagnostics up front
One of the usability issues with the new grouped diagnostics code is
that the actual error/warning/remark buried within the source code,
with only a header like

    === t.swift:1 ===

to show where the issue occurs. Instead, display the diagnostic
file:line:column and text right up front, the same way as the existing
compiler's style, e.g.,

    t.swift:5:3: error: no exact matches in call to global function 'f'

The is better for humans, because it puts the primary issue right up
front, and also better for tools that scrape logs to find the
diagnostics using this form.