Use Trivia(joining:) in SyntaxCollections
We've introduced `Trivia(joining: [Trivia], separator: [Trivia])` to
generate doc comments from multiple pieces, and already use this
approach in `SyntaxNodesFile` for nodes themselves, and for node
children documentation.

This commit switches from joining pieces manually to using
`Trivia(joining:)` in `SyntaxCollectionsFile` so that we use the same
approach everywhere.