Improve assertMacroExpansion highlight verification
Adds a new initializer to DiagnosticSpec which takes an optional array
of highlights instead of a single optional highlight string. Updates
existing initializer to map the string into an array of one.

Updates assertMacroExpansion to check each highlight in the diagnostic
spec against the actual produced diagnostic individually. Additionally
changes the behavior to drop leading and trailing trivial when
performing the highlight comparison to better match the Swift compiler's
behavior.