Add "lexical context" information to the macro expansion context
The lexical context of a particular macro expansion involves all of the
enclosing contexts, including functions, types and extensions,
properties and subscripts, and so on. The lexical context stack can be
used to, for example, determine which type a macro was expanded
within, or gather the parameters of the enclosing function for an
assertion, etc.