Add representation and parser support for the @isolated type attribute
As part of this, I had to update the attribute parser to not recurse into
the full parseType routine since that would attempt to parse attributes
again right after the `@`.  Doing so allows some genuinely bizarre constructs
that we obviously don't want and will reject in semantic analysis anyway.
For `@isolated` specifically, it creates an artificial parsing conflict
with the `isolated` parameter specifier and possible future declaration
modifier, just in case you wanted attributes to apply to your attributes.