Use presumed file/line number when printing grouped diagnostics
For the primary diagnostic, print the presumed file/line number in the
same manner as the Swift compiler does. In the actual code listing,
keep the line numbers and code from the actual source file that was
parsed, because there's no sensible way to map the source from the
`#sourceLocation`-named file (which might not exist).