Parse `#if canImport` as a function call instead of a specialized expression node
To make it easier to interpret the version passed to `canImport(MyModule, _version: <#version#>)`, add `ExprSyntax.interpretedAsVersionTuple` that takes eg. `1.2.3`, which got parsed as a member access `3` to the `1.2` float literal and re-parses it as a `VersionTupleSyntax`.

This relaxation also allows string literals as version parameters to `#if canImport`.

Fixes #1972
rdar://121070235

Co-Authored-By: Ziyang Huang <stevenhuang12@outlook.com>