Reimplement `SyntaxProtocol.isActive(in:)` in terms of ConfiguredRegions.
The compiler's semantics for determining unparsed vs. inactive regions
require looking all the way up the `#if` tree to the root, and the
bottom-up implementation of `isActive(in:)` was failing to match the
semantics of configured regions. Use `ConfiguredRegions` directly, with
a note that the performance of `SyntaxProtocol.isActive(in:)` is not
good.

I'm still considering whether this API should go away entirely.