Small improvements for ConfiguredRegions
Add a debugDescription so it's easier to see what the regions are, and allow
us to test for exact results.

Tweak the set of configured regions more precisely so they match what the
compiler is expecting. The only difference here is that we create a separate
region for an active block within another active block.