Teach the ActiveSyntaxRewriter to use either ConfiguredRegions or BuildConfiguration
Instead of always forcing the construction of ConfiguredRegions ahead of
time, introduce the notion of an ActiveClauseEvaluator that stores
either configured regions or a build configuration, and can query
for the active clause within a given #if declaration from either
representation. Use this in ActiveSyntaxRewriter to make the operation
itself single-pass, always.

Always test both paths to make sure they stay in sync.