Teach ActiveSyntax(Any)Visitor to use configured regions or a build configuration
Rework the implementations of ActiveSyntax(Any)Visitor to allow configured
regions, so that clients can re-use the work of evaluating the build
configuration for later traversals.

As a side effect of this, ActiveSyntax(Any)Visitor are no longer generic
over the type of build configuration, which should make them easier to
use.