Use fully-qualified types instead of `@retroactive` conformances
When increasing the tools version of swift-syntax to 5.9 or higher, the package is built with `-package-name`. It appears that conformances within the same package domain are not considered retroactive, which means that the addition of `@retroactive` produces a compilation error.

To allow swift-syntax to build both within a single package domain and without, use fully-qualified types instead of `@retroactive` in the potentially retroactive conformances.

Fixes #2844