Move contributor documentation into a dedicated folder
We had documentation targeted at swift-syntax contributors spread in the docc documentation of SwiftSyntax and SwiftParser. I think it doesn’t belong there because users of swift-syntax don’t care about it. It also makes it harder to find because you need to look in two places.

Move all that documentation to a single `Contributor Documentation` folder.